refactor(debug): extract pure utilities to debugHelpers.ts
Moves safeStringify, formatArgs, escapeRegExp, parseTimeValue,
parseLogSource, startCase, summarizeValue, buildMeta, buildDetails,
formatHeadline, trimSentence, structureLogEntry, canGroupStructuredRows,
groupStructuredRows out of DebugView.tsx into a sibling helpers file.

Also moves the LEVELS / LOG_WINDOW_STEP / SEARCH_LOG_WINDOW constants
and the LogSource / StructuredLogRow / LevelCounts / LevelFilter /
SummaryTone / DebugSnapshot types as named exports.

Pure refactor — no behavior change, no rendering change. Prep for
the sub-component split (T02-T03) where these helpers will be
consumed across multiple files.
